+++ demos/browser/app/meetingV2/meetingV2.ts
@@ -33,6 +33,8 @@ import {
   Versioning,
   VideoTileState,
   ClientVideoStreamReceivingReport,
+  SimulcastLayers,
+  SimulcastUplinkObserver
 } from '../../../../src/index';
 
 class DemoTileOrganizer {
@@ -110,7 +112,20 @@ export enum ContentShareType {
   VideoFile,
 };
 
-export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver, ContentShareObserver {
+const SimulcastLayerMapping = {
+  [SimulcastLayers.Low]: 'Low',
+  [SimulcastLayers.LowAndMedium]: 'Low and Medium',
+  [SimulcastLayers.LowAndHigh]: 'Low and High',
+  [SimulcastLayers.Medium]: 'Medium',
+  [SimulcastLayers.MediumAndHigh]: 'Medium and High',
+  [SimulcastLayers.High]: 'High'
+};
+
+export class DemoMeetingApp implements
+  AudioVideoObserver,
+  DeviceChangeObserver,
+  ContentShareObserver,
+  SimulcastUplinkObserver {
   static readonly DID: string = '+17035550122';
   static readonly BASE_URL: string = [location.protocol, '//', location.host, location.pathname.replace(/\/*$/, '/').replace('/v2', '')].join('');
   static testVideo: string = 'https://upload.wikimedia.org/wikipedia/commons/transcoded/c/c0/Big_Buck_Bunny_4K.webm/Big_Buck_Bunny_4K.webm.360p.vp9.webm';
@@ -1368,6 +1383,9 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
 
   audioVideoDidStart(): void {
     this.log('session started');
+    if (this.enableSimulcast) {
+      this.audioVideo.addSimulcastUplinkPolicyObserver(this);
+    }
   }
 
   audioVideoDidStop(sessionStatus: MeetingSessionStatus): void {
@@ -1580,6 +1598,14 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
 
   contentShareDidUnpause(): void {
     this.log(`content share unpaused.`);
+  }
+
+  encodingSimulcastLayerDidChange(simulcastLayers: SimulcastLayers): void {
+    this.log(`current active simulcast layers changed to: ${SimulcastLayerMapping[simulcastLayers]}`);
+  }
+
+  remoteVideosAvailableDidChange(attendees: { attendeeId: string; externalUserId: string }[]) {
+    this.log(`available remote videos changed: ${JSON.stringify(attendees)}`);
   }
 }